Removing FimImageKind
The discriminated unions feature of TypeScript is causing more trouble
than it was worth. Although the switch statements were nicer than a
bunch of if-instanceof statements, the type property had to be
initialized outside of the constructor, which forces the first statement
of the constructor to be a super() call.

This restriction was incompatible with the safety-checking for
width/height.

Also, these changes revealed a circular reference in the import
statements--to break it, files in the image/ and gl/ directories import
specific files, not index.ts of the other.