Changing performance measurements to use fixed blocks of time
Previously, the code would measure the time of a certain number of
iterations. Occasionally, the estimated number of iterations would be
wildly high, causing the test to be extremely slow.

This seems to have caused further problems on the iPad however--the
canvas copy test now reliably reboots the entire iPad.