Adding Fim factory class
Eventually, all constructors will become protected, and the only way
to create canvases will be through this factory class. This will
enable us to easily swap out implementations (for browsers vs Node).

This commit simply introduces the new factory class and the plumbing
to keep a member variable everywhere--it's not actually used yet.