Adding interfaces and internal classes to all remaining classes
Originally, my thought was that only FimCanvas and FimGLCanvas should
use the factory instead of constructors. However, the copyFrom() and
copyTo() methods of even the buffer classes can have some browser-
and WebGL-specific code in them. Easiest to just make everything use
the factory. The downside of this is that minimizers won't be able to
do tree shaking as well to eliminate dead code from parts of the FIM
library that aren't used.