Refactoring to separate parent and root objects
Originally, FIM v2 was designed around a single-level tree.
However, the introduction of 'operations' changed that, breaking
previous assumptions in the code.

Separating the two and using TypeScript to catch usage of the wrong
one. Also, in doing so, created FimObjectBase, an abstract class
which combines the functionality of EngineObject and FimOperation.