WIP: Refactoring JPEG/PNG handling
The intent here was to create a set of CoreImageFile classes to
do the JPEG/PNG import/export.

It didn't go as well as initially expected--there's likely a net
increase in code.

It also showed that on the most common FimEngine.createFromXYZ(),
we're decoding the image twice, which should be fixed.

There were also quite a few typos in JPEG vs. PNG, which may have
caused bugs in some cases.