Removing createImageWithFillColorAsync()
Specifing defaultFillColor in the image options is better in 90%
of use cases, as it covers cases like WebGL context loss too. It
also doesn't require an async call, so it can be called from a
constructor. If a caller really needs the old behavior, just call
createImage() followed by fillSolidAsync().