Making all textures default to 8 BPP until they are written to
When copying from a JPEG, PNG, or other 8 BPP source, it's a waste
of GPU memory to store the texture in a higher color depth. Rather
than making the client application pick the correct color depth,
we'll automatically use the lower and scale up to higher when its
actually needed.